Extract AzureServiceClient.getLongRunningOperationResult() to HttpLongRunningOperationResult
